requirements.txt
pandas
numpy
sqlalchemy
psycopg[binary]
python-dotenv
asyncpg
psycopg2